facebook_wrapper_code/facebook.py
import os
from flask import Flask, jsonify, request
from facebook_business.api import FacebookAdsApi
from facebook_business.adobjects.adaccount import AdAccount
from google.cloud import bigquery
import logging
import requests
from concurrent.futures import ThreadPoolExecutor, as_completed

# Configure Logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Constant Facebook App Credentials
MY_APP_ID = os.getenv('MY_APP_ID')
MY_APP_SECRET = os.getenv('MY_APP_SECRET')

# Initialize BigQuery client
bq_client = bigquery.Client()

app = Flask(__name__)

def status_update(jobId: str, status: str):
    """Update job status in an external job tracker."""
    url = "https://your-api-url.com"  # Update with actual API URL
    headers = {'Content-Type': 'application/json'}
    data = {'jobId': jobId, 'status': status}
    try:
        response = requests.post(url, json=data, headers=headers)
        response.raise_for_status()  # Raise error for bad response
    except requests.exceptions.RequestException as e:
        logger.error(f"Error updating job status: {e}")

def fetch_facebook_data(params):
    """Fetch Facebook Ads data."""
    # Fetching logic here
    pass

def process_and_upload_data(data, table_name):
    """Process and upload data to BigQuery."""
    # Processing and uploading logic here
    pass

@app.route('/facebook-data', methods=['POST'])
def facebook_data():
    """Cloud Function entry point."""
    try:
        if request.content_type != 'application/json':
            error_msg = 'Unsupported Media Type. Expected application/json.'
            logger.error(error_msg)
            return jsonify({'success': False, 'error': error_msg}), 415

        params = request.get_json()
        job_id = params.get('jobId', 'unknown')

        # Update job status to 'inProgress'
        status_update(job_id, 'inProgress')

        # Fetch and process data
        data = fetch_facebook_data(params)

        # Upload to BigQuery
        if 'table_name' in params:
            process_and_upload_data(data, params['table_name'])

        # Update job status to 'success'
        status_update(job_id, 'success')

        # Return success response with processed data
        return jsonify({'success': True, 'data': data}), 200

    except Exception as e:
        logger.exception('An error occurred.')
        status_update(job_id, 'failed')
        return jsonify({'success': False, 'error': str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True)
