facebook_wrapper_code/requirements.txt
Flask==2.0.1
google-cloud-bigquery==3.0.0
facebook-business==12.0.0
pandas==1.2.0
requests==2.25.0
